rlib library loading: deal with linker scripts in .so files
Some linux distros (Arch, Gentoo) put linker scripts in .so files. GNU ld
parses them and loads the correct library from a different file during the
linking. Unfortunately this doesn't work when loading dynamically, but the
actual library path is easy to parse from the script.

When loading dynamic libraries, in case of a certain loading error, retry
loading the library assuming it is actually a linker script.