Remove objspace.usepycfiles option.
This option was needed for the sandbox feature, where it might not be allowed
to write to the file system.  Nowadays Python has a sys.dont_write_bytecode
flag, which can be used for that.  By default, this flags is `False`, unless
when the sandbox feature is enabled; in this case it's set to `True`.